:lipstick: 追加辞書UIの修正

+++ FlickSKK/AdditionalDictionaryViewController.swift
@@ -5,15 +5,8 @@ import UIKit
 //
 // 辞書の更新は、URLをどこに保持するかが難しいので、現バージョンは対応しない。
 class AdditionalDictionaryViewController: SafeTableViewController {
-    private var entries : [String] = []
-
-    private let dictionaries : [(title: String, url: String)] = [
-        (title: "人名辞書", url: "http://openlab.jp/skk/skk/dic/SKK-JISYO.jinmei"),
-        (title: "郵便番号辞書", url: "http://openlab.jp/skk/skk/dic/zipcode/SKK-JISYO.zipcode"),
-        (title: "沖縄辞書", url: "http://openlab.jp/skk/skk/dic/SKK-JISYO.okinawa"),
-        (title: "絵文字", url: "https://raw.githubusercontent.com/uasi/skk-emoji-jisyo/master/SKK-JISYO.emoji.utf8"),
-        (title: "その他(URL指定)", url: "")
-    ]
+    private var entries : [AdditionalDictionaries.Entry] = []
+    private var dictionaries : [AdditionalDictionaries.Entry] = []
 
     init() {
         super.init(style: .Grouped)
@@ -36,8 +29,10 @@ class AdditionalDictionaryViewController: SafeTableViewController {
     // MARK: - Entries
 
     private func reloadEntries() {
-        self.entries = SKKDictionary.additionalDictionaries()
-         self.tableView.reloadData()
+        let action = AdditionalDictionaries()
+        self.entries = action.enabledDictionaries()
+        self.dictionaries = action.availableDictionaries()
+        self.tableView.reloadData()
     }
 
     func applicationDidBecomeActive(notification: NSNotification) {
@@ -46,21 +41,25 @@ class AdditionalDictionaryViewController: SafeTableViewController {
 
     // MARK: - Table View
     func numberOfSectionsInTableView(tableView: UITableView) -> Int {
-        return 3 // entries + quickadd + link
+        return 2 // entries + quickadd + link
     }
 
     func tableView(tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
         switch section {
-        case 0: return NSLocalizedString("HowToAddDictionary", comment: "")
-        case 1: return NSLocalizedString("UsefulDictionary", comment: "")
-        case 2: return NSLocalizedString("LinkToOpenlab", comment: "")
+        case 0:
+            if self.entries.isEmpty {
+                return NSLocalizedString("No dictionaries", comment: "")
+            } else {
+                return NSLocalizedString("EnableDictionaries", comment: "")
+            }
+        case 1: return NSLocalizedString("AvailableDictionaries", comment: "")
         default: return nil
         }
     }
 
     func tableView(tableView: UITableView, titleForFooterInSection section: Int) -> String? {
         switch section {
-        case 0: return NSString(format: NSLocalizedString("%d dictionaries registered", comment: ""), self.entries.count) as String
+        case 0: return nil
         default: return nil
         }
     }
@@ -69,7 +68,6 @@ class AdditionalDictionaryViewController: SafeTableViewController {
         switch section {
         case 0: return self.entries.count
         case 1: return self.dictionaries.count
-        case 2: return 1
         default: return 0
         }
     }
@@ -81,11 +79,11 @@ class AdditionalDictionaryViewController: SafeTableViewController {
         switch indexPath.section {
         case 0:
             cell.selectionStyle = .None
-            cell.textLabel?.text = self.entries[indexPath.row].lastPathComponent
+            cell.textLabel?.text = self.entries[indexPath.row].title
+            cell.accessoryType = .None
         case 1:
             cell.textLabel?.text = self.dictionaries[indexPath.row].title
-        case 2:
-            cell.textLabel?.text = NSLocalizedString("OpenSKKDictWiki", comment: "")
+            cell.accessoryType = .DisclosureIndicator
         default:
             fatalError("section > 2 has not been implemented")
         }
@@ -102,8 +100,6 @@ class AdditionalDictionaryViewController: SafeTableViewController {
                 done: {
                     self.reloadEntries()
             }), animated: true)
-        case 2:
-            UIApplication.sharedApplication().openURL(NSURL(string: "http://openlab.ring.gr.jp/skk/wiki/wiki.cgi?page=SKK%BC%AD%BD%F1")!)
         default:
             ()
         }
@@ -111,9 +107,10 @@ class AdditionalDictionaryViewController: SafeTableViewController {
 
     func tableView(tableView: UITableView, commitEditingStyle editingStyle: UITableViewCellEditingStyle, forRowAtIndexPath indexPath: NSIndexPath) {
         if indexPath.section == 0 && editingStyle == .Delete {
-            let path = self.entries[indexPath.row]
-            NSFileManager.defaultManager().removeItemAtPath(path, error: nil)
-            self.reloadEntries()
+            if let path = self.entries[indexPath.row].path {
+                NSFileManager.defaultManager().removeItemAtPath(path, error: nil)
+                self.reloadEntries()
+            }
         }
     }
